+++ src/primitive/tbl/ft_tblrmline.c
@@ -6,25 +6,25 @@
 /*   By: rlutt <rlutt@student.42.us.org>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/20 12:22:57 by rlutt             #+#    #+#             */
-/*   Updated: 2017/07/27 19:24:30 by rlutt            ###   ########.fr       */
+/*   Updated: 2018/04/13 13:39:38 by dauie            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../incl/tbl.h"
 #include "../../../incl/printf.h"
 
-char			**ft_tblrmline(char **tbl, char *rm, int len)
+char			**ft_tblrmline(char **tbl, char *rm, size_t len)
 {
-	int		i;
-	char	**res;
-	char	**tmp;
+	int			i;
+	char		**res;
+	char		**tmp;
 
 	i = -1;
 	res = NULL;
 	if (!(res = (char **)ft_memalloc(sizeof(char *) * len)))
 		return (NULL);
 	tmp = res;
-	while (++i < len && tbl[i])
+	while (++i < (int)len && tbl[i])
 	{
 		if (ft_strncmp(tbl[i], rm, ft_strlen(rm)) != 0)
 		{
